Add files via upload
This dataset comprises of sales transactions captured at a retail store.The dataset has 550,069 rows and 12 columns.


User_ID: Unique ID of the user. There are a total of 5891 users in the dataset.
Product_ID: Unique ID of the product. There are a total of 3623 products in the dataset.
Gender: indicates the gender of the person making the transaction.
Age: indicates the age group of the person making the transaction.
Occupation: shows the occupation of the user, already labeled with numbers 0 to 20.
City_Category: User's living city category. Cities are categorized into 3 different categories 'A', 'B' and 'C'.
Stay_In_Current_City_Years: Indicates how long the users has lived in this city.
Marital_Status: is 0 if the user is not married and 1 otherwise.
Product_Category_1 to _3: Category of the product. All 3 are already labaled with numbers.
Purchase: Purchase amount.